Added: binary_tree folder.
The new addition to the python magic repository is my implementation of
the binary tree. Hopefully I will find the time to expand it and
implement other kinds of structures. The current binary tree is simple
yet functional and can be used as is. But obviously the purpose of me
writing it lies in the educational value more than in the practical.
There are much better trees out there.